% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{comparison_diagnostics}
\alias{comparison_diagnostics}
\title{Emulator Diagnostic Plot}
\usage{
comparison_diagnostics(
  emulator,
  input_points,
  output_points,
  range,
  sd = 3,
  output_name
)
}
\arguments{
\item{emulator}{An \code{\link{Emulator}} object.}

\item{input_points}{A set of validation points.}

\item{output_points}{The outputs, \code{f(x)}, from the simulator.}

\item{range}{The expected range of the output.}

\item{sd}{Numeric: the allowed number of standard deviations (default: 3).}

\item{output_name}{Optional. A name for the output.}
}
\value{
A list of points whose emulator value is outside the allowed standard deviation.
}
\description{
Produces a diagnostic plot of emulator output.
The emulator output \code{E[f(x)]} is plotted against the simulator output \code{f(x)},
with error bars given by the emulator standard deviation \code{sqrt(3*Var[f(x)])},
for each point \code{x} in a validation set \code{X}. Points whose emulator expectation
lies outside 3-sigma of the simulator output are shown in red, and those input points
are returned.
}
\examples{
    in_vars <- c("aSI", "aIR", "aSR")
    out_vars <- c("nS", "nI", "nR")
    c_lengths <- c(0.1, 0.085, 0.075)
    base_emulators <- emulator_from_data(GillespieSIR, in_vars,
     out_vars, c_lengths = c_lengths)
    trained_emulators <- purrr::map(seq_along(base_emulators),
     ~base_emulators[[.x]]$bayes_adjust(GillespieSIR[,in_vars], GillespieSIR[,out_vars[[.x]]]))
    comparison_diagnostics(trained_emulators[[1]], GillespieSIR[,in_vars],
     GillespieSIR[,'nS'], list(c(0,1000),c(0,1000)))
}
