% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulatorr.R
\docType{package}
\name{emulatorr}
\alias{emulatorr}
\title{emulatorr: An emulation and history matching package.}
\description{
A set of functions for Bayes Linear emulation and history matching. For details on the
mathematical background, there are many papers freely available on the topic; for details
of the functions in this package, consult the manual or the help files for individual
functions or objects.
}
