% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_wave.R
\name{full_wave}
\alias{full_wave}
\title{History Match}
\usage{
full_wave(
  input_data,
  validation_data,
  ranges,
  output_names,
  targets,
  n_points = 40,
  previous_wave = NULL,
  sample_method = "lhs",
  ...
)
}
\arguments{
\item{input_data}{The set of training points}

\item{validation_data}{The set of points to use in validation}

\item{ranges}{The ranges of the inputs, as a named list.}

\item{output_names}{The names of the outputs to emulate.}

\item{targets}{The observations, given in the usual \code{(val, sigma)} form}

\item{n_points}{The number of points to evaluate the parameters on.}

\item{previous_wave}{The preliminary emulators for the set of waves, if they exist}

\item{sample_method}{The method to be used to find new points (see \code{\link{generate_new_runs}})}

\item{...}{Any optional parameters to pass to \code{\link{emulator_from_data}}}
}
\value{
A list of base emulators, trained emulators for this wave, new sample points, and new ranges.
}
\description{
Performs a full wave of emulation and history matching from data.
Given simulator runs (split into training and validation data), the target values,
and the identification of outputs to emulate, the function generates trained emulators,
tests them with emulator diagnostics (removing any emulators whose outputs cannot be
well emulated from the data), and finally generates a new sample of points to be entered
into the simulator.
}
\details{
Necessary parameters to be passed are the input data, the validation data, the ranges of
inputs, and the observation values for each output. If any specifications are to be passed
directly to the emulator construction, then they should be given as additional parameters
(see \code{\link{emulator_from_data}} to see the options).

If the wave to be emulated is the first wave, then a set of preliminary ('wave 0') emulators
are fitted to the data. The proper set of emulators are then trained using Bayes Linear
adjustment. On subsequent waves, the preliminary emulators should be passed to the function
as the argument \code{previous_wave}.

The output consists of a list of four items: the preliminary emulators \code{base_emulators},
the trained emulators \code{emulators}, the next points to be put into the simulator
\code{next_sample}, and the minimum enclosing hyperrectangle for the non-implausible
region, given as ranges \code{new_ranges}.
}
\examples{
 #ranges <- list(aSI = c(0.1, 0.8), aIR = c(0, 0.5), aSR = c(0, 0.05))
 #outputs <- c('nS','nI','nR')
 #targets <- list(
 # list(val = 281, sigma = 10.43),
 # list(val = 30, sigma = 11.16),
 # list(val = 689, sigma = 14.32)
 #)
 #wave1 <- full_wave(GillespieSIR, GillespieValidation, ranges, outputs, targets,
 # n_points = 30, deltas = rep(0.1, 3), quadratic = TRUE)
}
