% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelbuilding.R
\name{get_coefficient_model}
\alias{get_coefficient_model}
\title{Model Generation}
\usage{
get_coefficient_model(data, ranges, output_name, add = FALSE, order = 2)
}
\arguments{
\item{data}{A \code{data.frame} containing the input and output values}

\item{ranges}{A named \code{list} consisting of the ranges of the input parameters}

\item{output_name}{A string corresponding to the output to be modelled}

\item{add}{Should we perform stepwise add or stepwise delete? Default: \code{FALSE}}

\item{order}{To what order terms should the model be fitted? Default: 2 (quadratic)}
}
\value{
The fitted model
}
\description{
Creates a best fit of coefficients for a given data set.
}
\details{
There are two ways to generate the model; either start with all possible terms
(including cross-terms) up to order \code{n}, and then stepwise remove them;
or start with an intercept and stepwise add terms up to order \code{n}, only
retaining a term if the Information Criterion is improved. Which method is
chosen is dependent on the value of \code{add}; in the event where
\code{add = FALSE} and there are not enough degrees of freedom to start with
all possible terms, a warning will be given.
}
