% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelbuilding.R
\name{get_linear_model}
\alias{get_linear_model}
\title{Linear Model Generation}
\usage{
get_linear_model(data, ranges, output_name, add = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} containing the input and output values}

\item{ranges}{A named \code{list} consisting of the ranges of the input parameters}

\item{output_name}{A string corresponding to the output to be modelled}

\item{add}{Should we perform stepwise add or stepwise delete? Default: \code{FALSE}}
}
\value{
The fitted model
}
\description{
Creates a linear model from data and a list of inputs.
There are two ways to generate the model; either start with all possible linear
terms, and then stepwise remove them (using \code{step}), or start with an intercept
add stepwise add linear terms, only retaining them if the AIC is improved. Which is
chosen is dependent on the value of \code{add}; in the event where \code{add = FALSE}
and there are not enough degrees of freedom, a warning will be given.
}
