% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelbuilding.R
\name{hyperparam_fit}
\alias{hyperparam_fit}
\title{Hyperparameter Estimation}
\usage{
hyperparam_fit(inputs, model)
}
\arguments{
\item{inputs}{The data to fit to}

\item{model}{The pre-fitted model (determined using \code{step})}
}
\value{
A list of hyperparameter values.
}
\description{
Does hyperparameter fitting using the nlme package
}
