% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{output_plot}
\alias{output_plot}
\title{Emulator Plots with Outputs}
\usage{
output_plot(emulators, targets, points = NULL, npoints = 1000)
}
\arguments{
\item{emulators}{A list of \code{\link{Emulator}} objects}

\item{targets}{A named list of observations, given in the usual form}

\item{points}{A list of points at which the emulators should be evaluated. Default: \code{NULL}}

\item{npoints}{If no points provided, how many input points to evaluate? Default: 1000}
}
\value{
A ggplot object.
}
\description{
Plots emulator outputs across a set of points, with the corresponding observations
overlaid (with the appropriate uncertainty).
If no points are provided, then \code{npoints} points are uniformly sampled from the
input region. Else the provided points are used - for example, if a non-implausible
space is known.
}
