% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_implausible}
\alias{plot_implausible}
\title{Implausibility Plots}
\usage{
plot_implausible(df, names, nvars, min = FALSE, ticks = FALSE)
}
\arguments{
\item{df}{A data.frame containing the inputs and implausibilities}

\item{names}{A list of names for the inputs.}

\item{nvars}{The number of variables in the plotting region: either 1 or 2.}

\item{min}{Should minimum implausibility be plotted? Default: FALSE}

\item{ticks}{Are axis ticks required? Default: FALSE}
}
\value{
A ggplot object
}
\description{
Generates implausibility plots: either optical depth, or minimum implausibility.
Simply a plot wrapper to make later things easier for lattice plots, etc.
However, it does have value in and of itself.
}
