% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_lattice}
\alias{plot_lattice}
\title{Implausibility Plot Lattice}
\usage{
plot_lattice(implausibilities, targets, ranges, np = 50, ...)
}
\arguments{
\item{implausibilities}{A list of implausibilities for input points}

\item{targets}{A list of observations, given in the usual form}

\item{ranges}{The ranges of each of the input parameters.}

\item{np}{The number of bins in each part of the plot grids}

\item{...}{Any additional inputs (to be passed to \code{\link{optical_depth}})}
}
\value{
A ggplot object
}
\description{
Creates a lattice of plots, consisting of optical depths (both univariate and
two-variable projections), and minimum implausibility.
It is advisable to pass no more than 6 input parameter names to this function,
as the graphs will be too small for meaningful analysis if more are provided.
}
