% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{validation_diagnostics}
\alias{validation_diagnostics}
\title{Emulator Diagnostics}
\usage{
validation_diagnostics(
  emulators,
  validation_points,
  output_names,
  which_diag = "all",
  targets = NULL,
  ...
)
}
\arguments{
\item{emulators}{A list of emulators on which to perform diagnostics}

\item{validation_points}{The validation set}

\item{output_names}{The list of outputs to perform diagnostics on}

\item{which_diag}{Which diagnostics should be performed?}

\item{targets}{If required, the list of observations for the outputs}

\item{...}{Any additional parameters to pass to the diagnostic tests.}
}
\value{
A data.frame containing the points that failed one or more diagnostic tests.
}
\description{
Plots standard diagnostics for emulators.
}
\details{
These diagnostics are based on having two datasets: a training set and a
validation set. The emulators will have been trained on the training set,
and the validation set is passed to the functions in this wrapper.

The current options for diagnostics (with the codes for which_diag) are:

  Standard Errors (se)

  Comparison Diagnostics (cd)

  Classification Error (ce)

  All of the above (all)

For details on each of these, see the help files for \code{\link{standard_errors}},
\code{\link{comparison_diagnostics}} and \code{\link{classification_error}} respectively.
}
\examples{
output_names <- c('nS','nI','nR')
ems <- emulator_from_data(GillespieSIR, output_names,
 ranges = list(aSI = c(0.1, 0.8), aIR = c(0, 0.5), aSR = c(0, 0.05)),
 quadratic = TRUE)
targets <- list(
 list(val = 281, sigma = 10.43),
 list(val = 30, sigma = 11.16),
 list(val = 689, sigma = 14.32)
)
validation_diagnostics(ems, GillespieValidation, output_names, targets = targets)
validation_diagnostics(ems, GillespieValidation, output_names, c('se','cd'))
validation_diagnostics(ems[1:2], GillespieValidation, output_names[1:2], 'ce', targets[1:2])
validation_diagnostics(ems, GillespieValidation, output_names,
 targets = targets, sd = 1, cutoff = 4)

}
