% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{validation_plots}
\alias{validation_plots}
\title{Validation Set Plotting}
\usage{
validation_plots(emulators, input_points, output_names)
}
\arguments{
\item{emulators}{A set of \code{\link{Emulator}} objects.}

\item{input_points}{The validation input points on which to evaluate the emulators}

\item{output_names}{The names of the output parameters.}
}
\description{
Plots each of the emulator outputs against each input. For each input parameter, the
emulator expectation is plotted for each output. These plots are presented as a set, to
better identify trends and dependencies in the emulators outputs.
}
\examples{
ems <- emulator_from_data(GillespieSIR, output_names = c('nS', 'nI', 'nR'),
 ranges = list(aSI = c(0.1, 0.8), aIR = c(0, 0.5), aSR = c(0, 0.05)),
 quadratic = TRUE)
validation_plots(ems, GillespieValidation[,1:3], c('nS', 'nI', 'nR'))
}
