% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelbuilding.R
\name{variance_emulator}
\alias{variance_emulator}
\title{Variance Emulator builder}
\usage{
variance_emulator(
  input_data_var,
  input_data_exp,
  npoints,
  output_names,
  ranges,
  kurt = 3,
  input_names = names(ranges),
  beta,
  u,
  c_lengths,
  funcs,
  bucov,
  deltas,
  ev,
  quadratic = TRUE,
  beta.var = FALSE,
  lik.method = "my"
)
}
\arguments{
\item{input_data_var}{Required. A \code{data.frame} containing the input parameters and output
variances from a set of simulator runs.}

\item{input_data_exp}{Required. A \code{data.frame} containing the input parameters and output
means from a set of simulator runs.}

\item{npoints}{The number of runs performed per observed point.}

\item{output_names}{Required. The list of outputs to emulate from \code{input_data}.}

\item{ranges}{A named list of parameter ranges.}

\item{kurt}{The value of the kurtosis. Default: 3}

\item{input_names}{A list of input_names (if \code{ranges} is not provided).}

\item{beta}{Optional: specifications for the regression coefficients, given as a list
of lists \code{list(mu, sigma)} (a la \code{\link{Emulator}} specification).}

\item{u}{Optional: the correlation structure for each output, given as a list of
lists \code{list(mu, sigma, corr)}.}

\item{c_lengths}{Optional: a set of correlation lengths.}

\item{funcs}{Optional: basis functions for the regression surface.}

\item{bucov}{Optional: a list of functions giving the covariance between each of the beta
parameters and u(x).}

\item{deltas}{Optional: the nugget terms to include in u(x).}

\item{ev}{Optional. Used for determining nugget terms in absence on delta}

\item{quadratic}{Optional: should the regression surface be linear or quadratic? Default: F}

\item{beta.var}{Optional: should the beta coefficient be assumed to be known or should model variance be included?}

\item{lik.method}{Optional: method used to determine hyperparameters sigma and theta.}
}
\value{
A list of \code{Emulator} objects.
}
\description{
Creates an emulator whose variance itself is emulated, for stochastic systems.
}
