% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{visualisation_plot}
\alias{visualisation_plot}
\title{Expectation and Variance Visualisation}
\usage{
visualisation_plot(ems, input_points, output_names)
}
\arguments{
\item{ems}{The list of emulators}

\item{input_points}{The points on which to evaluate the emulators}

\item{output_names}{The list of outputs}
}
\description{
Plots the points of a wave as a pairs plot, coloured by emulator expectation (lower) and
variance (upper) for each output.
}
\examples{
ranges <- list(aSI = c(0.1, 0.8), aIR = c(0, 0.5), aSR = c(0, 0.05))
targets <- list(
 list(val = 281, sigma = 10.43),
 list(val = 30, sigma = 11.16),
 list(val = 689, sigma = 14.32)
)
outputs <- c('nS','nI','nR')
ems <- emulator_from_data(GillespieSIR, outputs, ranges, deltas = rep(0.1, 3))
t_ems <- purrr::map(seq_along(ems), ~ems[[.]]$adjust(GillespieSIR, outputs[[.]]))
visualisation_plot(t_ems, GillespieSIR, outputs)

}
