% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{wave_points}
\alias{wave_points}
\title{Plot Points of Waves}
\usage{
wave_points(pts_list, in_names)
}
\arguments{
\item{pts_list}{A list object, whose elements are data.frames of points}

\item{in_names}{The input dimension names.}
}
\value{
A ggplot object.
}
\description{
Creates a set of pairs plots of the input points for multiple waves.
}
\details{
For each pair of inputs, the points from a succession of waves are plotted,
coloured according to the wave number. Histograms of the points for each input
(broadly interpreted as marginal distributions for the inputs) are provided
on the main diagonal.
}
