% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{comparison_diagnostics}
\alias{comparison_diagnostics}
\title{Emulator Diagnostic Plot}
\usage{
comparison_diagnostics(
  emulator,
  input_points,
  output_points,
  output_name,
  sd = 3,
  plt = T,
  ...
)
}
\arguments{
\item{emulator}{An \code{\link{Emulator}} object.}

\item{input_points}{A set of validation points.}

\item{output_points}{The outputs, \code{f(x)}, from the simulator.}

\item{output_name}{Optional. A name for the output.}

\item{sd}{Numeric: the allowed number of standard deviations (default: 3).}

\item{plt}{Should a plot be shown (default: T).}

\item{...}{Dummy parameters (for compatibility with diagnostic wrapper)}
}
\value{
A list of points whose emulator value is outside the allowed standard deviation.
}
\description{
Produces a diagnostic plot of emulator output.
}
\details{
The emulator output \code{E[f(x)]} is plotted against the simulator output \code{f(x)},
with error bars given by the emulator standard deviation \code{sqrt(3*Var[f(x)])},
for each point \code{x} in a validation set \code{X}. Points whose emulator expectation
lies outside 3-sigma of the simulator output are shown in red, and those input points
are returned.
}
\examples{
em <- emulator_from_data(GillespieSIR, output_names = c('nS', 'nI', 'nR'),
 ranges = list(aSI = c(0.1, 0.8), aIR = c(0, 0.5), aSR = c(0, 0.05)),
 quadratic = TRUE)[[1]]
comparison_diagnostics(em, GillespieValidation[,1:3], GillespieValidation[,'nS'])
#> (0.7864384, 0.01426296, 0.001072935)
}
