% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nthimplausible.R
\name{nth_implausible}
\alias{nth_implausible}
\title{n-th Maximum Implausibility}
\usage{
nth_implausible(emulators, x, z, n = floor(length(emulators)/2), max_imp = 20)
}
\arguments{
\item{emulators}{A set of \code{\link{Emulator}} objects.}

\item{x}{An input point}

\item{z}{The observed outputs, either as a numeric vector or as a collection of val, sigma pairs (see examples)}

\item{n}{The implausibility level to return. By default, the median implausibility is chosen}

\item{max_imp}{A maximum implausibility to consider: in most cases, it is useful to truncate the size of the I(x). Default: 20.}
}
\value{
The n-th maximum implausibility value.
}
\description{
For a collection of emulators, it can be helpful to combine the implausibility
measures for a given observation. The maximum implausibility is, simply, the largest
implausibility value given by the emulators for each output; the 2nd maximum is the maximum
of the set without the maximum, and so on.
}
\examples{
    in_vars <- c("aSI", "aIR", "aSR")
    out_vars <- c("nS", "nI", "nR")
    c_lengths <- c(0.1, 0.085, 0.075)
    base_emulators <- emulator_from_data(GillespieSIR, in_vars,
     out_vars, c_lengths = c_lengths)
    trained_emulators <- purrr::map(seq_along(base_emulators),
     ~base_emulators[[.x]]$bayes_adjust(GillespieSIR[,in_vars], GillespieSIR[,out_vars[[.x]]]))
   target_vals <- c(281, 30, 689)
   target_sigmas <- c(37.26, 11.16, 31.72)
   z_specs <- purrr::map2(target_vals, target_sigmas, ~list(val=.x, sigma=.y))
   nth_implausible(emulators = trained_emulators, x = c(0.4, 0.25, 0.025),
    z = z_specs, n = 2)
}
