% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{output_plot}
\alias{output_plot}
\title{Emulator Plots with Outputs}
\usage{
output_plot(emulators, targets, points = NULL, npoints = 1000)
}
\arguments{
\item{emulators}{A list of \code{\link{Emulator}} objects}

\item{targets}{A named list of observations, given in the usual form}

\item{points}{A list of points at which the emulators should be evaluated. Default: \code{NULL}}

\item{npoints}{If no points provided, how many input points to evaluate? Default: 1000}
}
\value{
A ggplot object.
}
\description{
Plots emulator outputs across a set of points, with the corresponding observations
overlaid (with the appropriate uncertainty).
If no points are provided, then \code{npoints} points are uniformly sampled from the
input region. Else the provided points are used - for example, if a non-implausible
space is known.
}
\examples{
 ranges <- list(aSI = c(0.1, 0.8), aIR = c(0, 0.5), aSR = c(0, 0.05))
 outputs <- c('nS','nI','nR')
 targets <- list(
  nS = list(val = 281, sigma = 10.43),
  nI = list(val = 30, sigma = 11.16),
  nR = list(val = 689, sigma = 14.32)
 )
 ems <- emulator_from_data(GillespieSIR, outputs, ranges,
  deltas = rep(0.1, 3), quadratic = TRUE)
 t_ems <- purrr::map(seq_along(ems), ~ems[[.]]$adjust(GillespieSIR, outputs[.]))
 output_plot(t_ems, targets)
}
