% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_lattice}
\alias{plot_lattice}
\title{Create a Plot Lattice}
\usage{
plot_lattice(ems, targets, ppd = 40, cb = FALSE)
}
\arguments{
\item{ems}{The list of emulators}

\item{targets}{The corresponding list of targets for the emulators}

\item{ppd}{The number of points to sample per input dimension. The granularity should be
carefully considered for large parameter spaces. Default: 40}

\item{cb}{Should a colourblind-friendly palette be used for implausibility? Default: FALSE}
}
\value{
A ggplot object
}
\description{
Plots a number of emulator plots, all projections of the full-dimensional space.
}
\details{
The plots to be included are:

One dimensional optical depth plots (the proportion of points that are non-implausible)

Two dimensional optical depth plots

Two dimensional minimum implausibility plots

The 1d optical depth plots are situated on the diagonal, the 2d optical depth plots in the
upper triangular elements, and the minimum implausibility plots in the lower triangular
elements. To evaluate the quantities, a regular grid with \code{ppd} points per dimension
is created, and maximum implausibility is calculated over this grid.
}
