% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleinputs.R
\name{scale_input}
\alias{scale_input}
\title{Scale input parameter values}
\usage{
scale_input(x, r)
}
\arguments{
\item{x}{The point to scale}

\item{r}{The list of ranges}
}
\value{
The scaled point
}
\description{
Given ranges and a point x taken from an arbitrary range, scale it
to be in the range [-1,1]*n (n is the number of input dimensions).
}
\examples{
   x <- c(0.1, 0.5, 0.025)
   ranges <- list(c(0.1,0.8), c(0, 0.5), c(0, 0.05))
   scale_input(x, ranges)
}
