% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{space_removed}
\alias{space_removed}
\title{Space Removal}
\usage{
space_removed(
  emulators,
  validation_points,
  z,
  n_points = 10,
  u_mod = seq(0.8, 1.2, by = 0.1),
  intervals = seq(0, 10, length.out = 200),
  modified = "disc"
)
}
\arguments{
\item{emulators}{A set of \code{\link{Emulator}} objects.}

\item{validation_points}{The validation set used in this wave.}

\item{z}{The observations with which to match, as \code{list(val, sigma)} pairs.}

\item{n_points}{The number of points in each dimension of the grid.}

\item{u_mod}{The percentage differences in structural discrepancy to examine.}

\item{intervals}{The set of implausibility cut-offs to consider.}

\item{modified}{What parameter should be varied in the analysis?}
}
\value{
A list of two \code{data.frame}s, one for removed space and one for misclassifications.
}
\description{
Finds the proportion of space removed as a function of implausibility cut-off, and of structural
discrepancy, or changed variance.
}
\details{
The reduction in space is found by evaluating over a p^d regular grid, where p is chosen by
\code{n_points} and d is the dimension of the input space. Larger values of \code{n_points}
will give a more accurate reflection of removed space, at high computational cost. For the
purpose of quick diagnostics, \code{n_points = 5} is acceptable.

The parameter \code{modified} can take three options: 'disc' (default) corresponding to
model discrepancy, 'var' corresponding to emulator variance, or 'corr' corresponding to
correlation length. In the first case, the implausibilities are recalculated with the
original emulators; in the latter two cases, the emulators are re-trained with the new
specifications. For this reason, one should expect the 'var' and 'corr' options to be
more computationally intensive.

The returned output is a \code{data.frame} consisting of the percentage of space
removed at each cutoff value, for each modifed value of the varied parameter. The main
result, however, is the accompanying plot of this information.
}
\examples{
ranges <- list(aSI = c(0.1, 0.8), aIR = c(0, 0.5), aSR = c(0, 0.05))
targets <- list(
 list(val = 281, sigma = 10.43),
 list(val = 30, sigma = 11.16),
 list(val = 689, sigma = 14.32)
)
outputs <- c('nS','nI','nR')
ems <- emulator_from_data(GillespieSIR, outputs, ranges, deltas = rep(0.1, 3), quadratic = TRUE)
t_ems <- purrr::map(seq_along(ems), ~ems[[.]]$adjust(GillespieSIR, outputs[[.]]))
names(t_ems) <- outputs
removal <- space_removed(ems, GillespieValidation, targets,
 n_points = 5, u_mod = seq(0.75, 1.25, by = 0.25), intervals = seq(2, 6, by = 0.1))

}
