% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{space_removed}
\alias{space_removed}
\title{Space Removal}
\usage{
space_removed(
  emulators,
  validation_points,
  z,
  n_points = 10,
  u_mod = seq(0.8, 1.2, by = 0.1),
  intervals = seq(0, 10, length.out = 200)
)
}
\arguments{
\item{emulators}{A set of \code{\link{Emulator}} objects.}

\item{validation_points}{The validation set used in this wave.}

\item{z}{The observations with which to match, as \code{list(val, sigma)} pairs.}

\item{n_points}{The number of points in each dimension of the grid.}

\item{u_mod}{The percentage differences in structural discrepancy to examine.}

\item{intervals}{The set of implausibility cut-offs to consider.}
}
\value{
A list of two \code{data.frame}s, one for removed space and one for misclassifications.
}
\description{
Finds the proportion of space removed as a function of implausibility cut-off, and of structural
discrepancy. Also shows the proportion of misclassifications given a chosen cut-off (applied to
both the simulator and emulator implausibilities).
}
\details{
The misclassifications are checked using the validation set from the current wave; the reduction
in space is found by evaluating over a p^d regular grid, where p is chosen by n_points and d is
the dimension of the input space.
}
