% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{standard_errors}
\alias{standard_errors}
\title{Emulator Standard Errors}
\usage{
standard_errors(emulator, input_points, output_points, output_name, plt = T)
}
\arguments{
\item{emulator}{An \code{\link{Emulator}} object.}

\item{input_points}{A set of validation points.}

\item{output_points}{The outputs, \code{f(x)}, from the simulator.}

\item{output_name}{Optional. A name for the output.}

\item{plt}{Should a plot be shown (default: T).}
}
\value{
A list of standard errors.
}
\description{
Finds and plots emulator standard errors.
For an emulator of a simulator function \code{f(x)}, and a validation data set \code{X},
finds the standard errors in the form \code{(f(x)-E[f(x)])/sqrt(Var[f(x)])},
where \code{E[f(x)]} is the emulator expectation, and \code{Var[f(x)]} is the emulator variance,
at each point \code{x} in \code{X}.
}
\examples{
em <- emulator_from_data(GillespieSIR, output_names = c('nS', 'nI', 'nR'),
 ranges = list(aSI = c(0.1, 0.8), aIR = c(0, 0.5), aSR = c(0, 0.05)),
 quadratic = TRUE)[[1]]
standard_errors(em, GillespieValidation[,1:3], GillespieValidation[,'nS'], 'nS')
#> (0.7864384, 0.01426296, 0.001072935)
}
